requirements.txt
fastapi==0.110.0
starlette==0.36.3
jinja2==3.1.2

uvicorn[standard]==0.29.0
gunicorn==21.2.0

sqlalchemy==2.0.29
psycopg2-binary==2.9.9

openai==1.30.1
requests==2.31.0

python-dotenv==1.0.1

python-telegram-bot==20.7

passlib[bcrypt]==1.7.4
bcrypt==3.2.2

itsdangerous==2.1.2
python-multipart==0.0.9

alembic==1.13.1

redis==5.0.4